client: Fix race condition in operation handler
As a goroutine is spawned for every event and we force queueing while we
get the initial state we need to make sure it's impossible for two
handlers to run concurently when touching shared resources.

We also need to cope with the fact that one of the handlers may have
decided to disconnect the listener and so should just return in such
case.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>