Complete cmd.Context support for various AskXXX methods
This branch is a mechanical follow-up of the first one introducing
cmd.Context. It adds the rest of AskXXX methods as defined in cmdInit,
and should be a step towards making cmdInit itself testable.

The logic is exactly the same as the inline functions currently
defined in main_init.go, although some boilerplate could be avoided by
factoring common logic together.

Signed-off-by: Free Ekanayaka <free@ekanayaka.io>