test: add LXD running Docker tests
This implements a minimal test for running Docker in LXD. The binaries are
always the ones from current Docker master which are built after a new commit
is pushed.

Currently the test only check if the daemon starts up correctly and whether an
image can be pulled.  Until Docker vendors a version of runc that includes a
fix for the non-dumpable bug there's really not a lot more for us to test.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>