Extract validation of --auto args into a separate method
This just mechanically moves into a standalone method the code that
validates the various arguments that go with --auto. It makes the rest
of the code a bit shorter and easier to refactor in the follow-up
commit.

For good measure, I've also added a few unit tests covering some of
the invalid arguments combinations or values.

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>