Extract code asking init questions to individual methods
As a step towards making CmdInit.runAutoOrInteractive() shorter and
easier to split, I mechanically extracted the code that asks for the
various configuration choices into standalone methods.

This makes both the higher-level logic more readable (i.e. it's now
immediate to check which questions get asked and in which order), and
isolates the lower-level logic of individual questions, so it's easier
to see and have confidence of what are the moving parts and variables
involved (without interference between the various questions).

The logic has been moved around without any change, I've just added an
"XXX" comment in the new askStorage() method since there are a couple
of checks which seem redundant and could be dropped.

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>