Fix regression in image auto-update logic
From what I understand, it seems that this commit:

https://github.com/lxc/lxd/commit/7d4869527f1cef579015a06258c93a33e76cc865

introduced a regression in the auto-update logic. The code that was
setting the AutoUpdate flag in the *api.Image structure in line 541:

https://github.com/lxc/lxd/commit/7d4869527f1cef579015a06258c93a33e76cc865#diff-6182827642cd00e061c6ffe52891acc5L541

was moved down in the function:

https://github.com/lxc/lxd/commit/7d4869527f1cef579015a06258c93a33e76cc865#diff-6182827642cd00e061c6ffe52891acc5R482

however the new location is after the call to dbImageInsert:

```
	err = dbImageInsert(d.db, info.Fingerprint, info.Filename, info.Size, info.Public, info.AutoUpdate, info.Architecture, info.CreatedAt, info.ExpiresAt, info.Properties)
```

that actually makes use of such setting. This results in the
auto_update column in the database always being 0, no matter what, and
the server logic to auto-update images will never kick in.

This commit should also fix another (non fatal) regression introduced
by:

https://github.com/lxc/lxd/commit/806d55c88562ab82b32d2d8563c20a8ec9bdac98

where appending the empty string "" to the poolNames to be processed
ends up triggering a call to ```doDeleteImageFromPool()``` from
``autoUpdateImage``:

https://github.com/lxc/lxd/blob/471604191b9793ffe479036dab8e527cb743cee1/lxd/images.go#L975

which as far as I understand will always end up with an error in case
poolName is the empty string, so a spurious error will be logged.

Note that auto-update regression affects the latest 2.13 release, but
not 2.12.

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>