If present, use cached images when launching containers
If the auto-update feature is on, prefer using cached images when
launching a container, even if the cache image is "stale" (i.e.
there is a more recent version of it in the given remote).

I've decided to add a new preferCached boolean parameter to
ImageDownload(), as it seemed the simplest option (and there are
already other boolean switches, like forContainer).

However, the API and implementation of ImageDownload() are getting a
bit fat/complicated, so at some point we might want to break it into
more modular pieces that offer simpler APIs.

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>