Don't special-case mock mode unnecessarily in db patches
A few db patches are currently checking if the daemon is being run in
mock mode (i.e. unit tests), and no-op the patch in that.

It turns out that this is not really necessary, becasue the underlying
logic will work fine in mock mode too (and also, more generally it
might open the doors to creating tests that specicically exercise
certain patches).

This change removes all current occurrences of the special casing, the
main goal would be to reduce coupling between daemon
code (higher-level) and db code (lower-level), so eventually we can
have dependencies only in one direction (deamon depends on db but not
viceversa).

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>