storage: add helper to detect if pool is in use
We should check whether a given storage entity that serves as a storage pool is
already in use by LXD if it already exists. For example, an already existing
volume group that is supposed to be used by LXD should be checked whether it is
already used by LXD. If so, error out. This will prevent a ton of bug reports.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>