Sync shared/api with master branch
This means that shared/api will now contain everything which was added
through API extensions after the release of LXD 2.0.

The LXD daemon in the 2.0 LTS branch will not be using any of those
added features, but having the API defintions in sync will allow for
keeping the client library in sync too (which is extension aware) and
overall make maintenance much simpler.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>