network: use lxc.network.<n>.*
This deprecates defining multiple networks in the following format:

	lxc.network.type = veth
	lxc.network.flags = up
	lxc.network.link = lxdbr0
	lxc.network.hwaddr = 00:16:3e:09:7a:39
	lxc.network.name = eth0

	lxc.network.type = veth
	lxc.network.flags = up
	lxc.network.link = lxdbr1
	lxc.network.hwaddr = 00:16:3e:b2:8f:56
	lxc.network.name = eth1

in favor of:

	lxc.network.0.type = veth
	lxc.network.0.flags = up
	lxc.network.0.link = lxdbr0
	lxc.network.0.hwaddr = 00:16:3e:09:7a:39
	lxc.network.0.name = eth0

	lxc.network.1.type = veth
	lxc.network.1.flags = up
	lxc.network.1.link = lxdbr1
	lxc.network.1.hwaddr = 00:16:3e:b2:8f:56
	lxc.network.1.name = eth1

The former syntax is utterly misleading and difficult to handle in liblxc while
the latter will let liblxc nicely handle network configuration even if you
specify everything in random order. Cleaning out and updating will also work
much more reliable.

The regression potential for LXD is basically 0 since we regenerate the config
for the container each time we start it up.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>