storage: allow attaching storage volumes
Since it's implementation the LXD storage API has supported attaching storage
volumes to containers. One of the oustanding challenges before a working
implementation of the shiftfs concept in the kernel is how to deal with id
mappings when mounting custom storage volumes. This spec tries to propose a
viable solution that aims at security.

When a new custom storage volume is created:

    lxc storage volume create <pool> <volume>

and attached to a container for the first time:

    lxc storage volume attach <pool> <volume> <container> <path-to-attach-to>

it will not have any prior id mapping associated with it.

1. Attaching a previously not id-mapped storage volume
The first container to attach the storage volume will perform an id mapping on
the storage volume. The serialized id mapping recorded in the containers config
file will then be stored as a volatile key in the storage volume's config.

2. Attaching an already id-mapped storage volume
When a storage volume is attached to a container it is checked whether the
volatile key is empty. If it is empty case 1. applies. If it is not empty it is
checked whether the id mapping of the container matched the idmap recorded in
the volatile key. If they match the container is allowed to attach the storage
volume. If they do not match it will be checked whether the storage volume is
atttached to any other containers. If it is attached to any other containers the
container is not allowed to attach the storage volume and an appriorate error is
returned. If it is not attached to any other containers the container is allowed
to attach the storage volume, map the ids and update the volatile key
appropriately.

3. Attaching a storage volume to multiple containers
When a storage volume is attached to multiple containers all containers must
have the same id mapping. This consensus must be upheld as long as multiple
containers share the same storage volume.

3. Changing the idmapping of a container with attached id-mapped storage volumes
If the idmapping of a container is changed that has an id-mapped storage volume
attached to it which is shared among multiple containers the change will not be
applied an approriate error is returned.

Closes #3389.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>