Implement instance types as proxy for limits
This makes it possible to use standard instance types as used by major
cloud providers to define LXD resource limits.

  lxc launch ubuntu:16.04 c1 -t t2.micro
  lxc launch ubuntu:16.04 c2 -t n1-highcpu-16
  lxc launch ubuntu:16.04 c3 -t Standard_A8_v2

LXD will attempt to find a unique match for the instance type.
If it's ambiguous, the cloud provider can be specified with:

  lxc launch ubuntu:16.04 c4 -t azure:A11

This can also be used with an internal format string to specify simple limits:

  lxc launch ubuntu:16.04 c5 -t c2.5-m2

    stgraber@castiana:~$ lxc config show c5 | grep limits
      limits.cpu: "3"
      limits.cpu.allowance: 83%
      limits.memory: 2048MB

All provided memory amounts are in GB, LXD will convert internally to MB
to allow for fractions. For fractions of CPU, LXD will apply both hard
CPU limits through pinning and a limit on CPU time to match the exact
provided limit when under load.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>