Move db*.go files into their own db/ sub-package
This branch is very large but almost completely mechanical. It moves
all db-related code into a standalone subpackage (db). This
facilitates testing and makes further refactoring easier to reason
about.

The only point worth mentioning is that containerArgs and
storagePoolVolumeTypeToName had to be moved to the db package, to
avoid circular dependencies. We will probably want to clean that up
down the road.

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>