Extract Daemon.SetupStorageDriver into a standalone function
Convert the Daemon.SetupStorageDriver method into a standalone
function accepting a Daemon parameter and move it to the storage.go
file. Later down the road it should be possible to make the function
accept a more narrow API as parameter than the whole Daemon structure,
for decoupling, easier dependency management and unit-testability.

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>