Add initial lxd/sys sub-package and OperatingSystem structure
Introduce an initial implementation of the sys.OS struct, which is
meant to be a high-level API to access system-related functionality.

For now it has a single attribute, OS.Architectures, to access the
available architectures (whose core generic logic has been moved to
the lxd/util sub-package).

The Daemon.architectures attribute has been dropped and replaced with
an instance of the OS struct. Further parts of Daemon will be factored
out down the road.

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>