Drop unnecessary checks on MockMode
Some code in the Daemon that was conditionally triggered if MockMode
was on, is now executed regardless of MockMode, since thre are already
checks on MockMode deeper in the APIs being executed (e.g. networkStartup)
or there's no need to skip the execution of those APIs, since they'll
work fine in our default unit tests setup (e.g. SetupStorageDriver).

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>