Complete moving schema creation logic to schema.Schema
This commit drops all remaining schema-related logic living directly
in lxd/db, since all of it can now be handled by lxd/db/schema.

Also, it adds a new db.Node struct which wraps the actual sql.DB
handle to the node-level database. This is a step torwards gradually
moving data and db APIs from the local node-level database to the
cluster one: the idea is that code consuming the lxd/db package won't
interact directly with sql.DB, but rather with the the higher-level
db.Node, whose APIs will be gradually migrated to a db.Cluster object
that will be introduced in follow-up commits.

Finally, initialization code that is specific to the nodel-level
database has been moved to a db/node sub-package.

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>