Rename Daemon.db to Daemon.nodeDB
This is a mechanical rename of the Daemon.db attribute to a temporary
Daemon.nodeDB. The plan is for Daemon.db to become an instance of the
high-level db.Node abstraction, instead of sql.DB, and for Daemon.nodeDB
to be dropped when the transition is completed. Later down the road
Daemon.db will be further migrated to db.Cluster, which will host most
of the APIs now on db.Node.

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>