Add initial Go-level daemon integration-like test
This commit introduces a newDaemon() testing function that can be used
to perform integration-like tests.

The idea is that the test daemon will run basically the same code as a
standalone compiled process, but should be more flexible and cheaper
to run for tests that need to exercise end-to-end behavior (such as
wiring code, etc).

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>