migration: restore state when moving container
When a container is running users can pass "--stateful" to the lxd command line
tool "lxc move". This will move the container and restore the state of the
container on the target.
My original implementation of this feature only applied to stateless container
copies but not to requests to move a container. The only thing we need to do is
pass the stateless argument along in the move codepath for the "lxc move"
command line tool and then restore state on the receiving side based on the
container's config. This is trivial since we know the source container's state
because we preserve volatile keys when moving containers.

Closes #3798.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>