ceph: handle volume.block.filesystem update
The ceph implementation for changing volume.block.filesystem the cherry-picked
commits is missing logic. ceph requires that the image is kept around when it
has dependent snapshots. So when you force an ImageDelete() on container
creation after a volume.block.filesystem update and the original container
still exists you mark the image as delete and the next ImageCreate() request
will restore the image that was simply marked as deleted which has the fs
before the volume.block.filesystem change. That's why the ceph part of this
patch still fails.

To clarify my point a little more concrete:

	lxc storage create pool ceph volume.block.filesystem=ext4
	  - creates "pool/image_<fp>" + "pool/image_<fp>@readonly"
	lxc launch images:alpine/edge a1 -s pool
	  - clones "pool/image_<fp>@readonly"

	lxc storage set pool volume.block.filesystem xfs
	lxc launch images:alpine/edge a2 -s pool
	  - zombifies "pool/image_<fp>" to "pool/zombie_image_<fp>"
	  - tries to create xfs version of image with fp <fp>
	    - checks for the existence of any zombified images
	      - will detect that "pool/zombie_image_<fp>" exists
	        - resurrects "pool/zombie_image_"fp">" to "pool/image_<fp>"
	  - clones new container from "pool/image_<fp>"
	  - calls xfs_admin for UUID generation
	    - fails because clone is ext4

My new logic changes the image image deletion routine and the container create
from image rountine to:

	ImageDelete()
	- if (!has_dependent_snapshots())
		wipe_image_from_disk()
	  else
	  	zombifie_image() {
			rename "pool/image_<fp>" to "pool/image_<fp>_<fstype>"
		}

	ContainerCreateFromImage()
	- if (!zombified_image_with_fstype_exists())
		resurrect()

Backwards compatibility:
This change is fully backwards compatible. Images that have been zombified
before will not be resurrected anymore since the new checks check for the
existence of the fs type suffix as well.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>