Add util.InMemoryNetwork to create in-memory listener/dialer pairs.
This is a convenience for creating in-memory networks that implement
the net.Conn interface. It will be used when running a node in
non-clustered mode, where there will be no actual TCP/gRCP connection
to an external dqlite node, but rather just an in-memory connection to
the local dqlite instance (which will be the leader).

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>