Control all goroutines spawned in Daemon.Ready() using task.Task
This reduces the boilerplate needed to handle updating the frequency
at which a task is executed, and provides a single entry point for
shutting down background tasks. Later down the road we'll want to add
support for graceful cancellation of tasks such the auto update images
one, which is particularly important for unit test isolation.

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>