Add a shared.KeyPairAndCA function to get coventionally named certs
The new KeyPairAndCA function plays well with the certificate naming
conventions and PKI support used in LXD and can be used as a
convenience to load/generate client or server certificates.

For easier testing of server code needing a server cert, two new
TestingKeyPair and TestingAltKeyPair functions, each one returning a
pre-canned CertInfo, have been added.

The initTLSConfig private function was exported, so we can easily
re-use our default TLS parameters wherever needed (basically once in
the client and once in the server, which at the moment duplicate those
settings).

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>