Add support for gracefully aborting schema.Ensure
This commits adds a new Check() method to the public API of
schema.Schema.

It can be used to perform some checks before actually triggering the
logic that ensures the schema is at the expected version. If the Check
function returns an error, then Ensure() will abort early. If the
error is the special value ErrGracefulAbort, the abort will be
graceful and any change performed by the Check function will be
committed.

The Check() method will be used for the cluster database, to check
that all nodes are at the same version, and possibly to update the
information of a specific node and say "I'm at this version, but I'll
won't update the schema since I see other nodes are behind".

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>