Extract the APIExtensions list from api10Get
The list of API extensions can now be accessed from other call
sites. This will be needed when figuring the "version" of an LXD
cluster node, which will be a combination of schema version and number
of supported API extensions (both of which increase monotonically).

The declaration has been put into the shared/version package for
consistency with the APIVersion variable already present.

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>