Gracefully cancel tasks on daemon shutdown
The various periodic tasks that the Daemon spawns have been changed so
they do their best to gracefully terminate as soon as possible when
the given context gets cancelled (which happens upn daemon
shutdown). Due to some of underlying APIs not supporting cancellation
yet, this should be considered just a starting point.

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>