client: add client API ConsoleContainer()
This add the ConsoleContainer() API extension. This enables the client to send
POST requests to LXD's /containers/<name>/console endpoint to attach to a
container's console device.

ConsoleContainer() takes a struct ContainerConsolePost as argument. In addition
it takes a struct ContainerConsoleArgs as argument which callers can use to
specify a control channel to send window resize requests to the daemon. It is
required that callers pass in a bidirectional file descriptor in order to
interact with the container's console device.

Closes #1129.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>