client: add client API GetContainerConsoleLog()
GetContainerConsoleLog() is modelled after GetContainerLogfiles() and returns
an io.ReadCloser. It takes a ContainerConsoleLogArgs struct as additional
argument. The struct is currently empty. However, once we hookup the ringbuffer
or if we want to expand this function in the future it is well worth it to have
this struct around.

Closes #1129.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>