lxd/dameon: Add LXD_EXEC_PATH to override execPath
This is needed for the LXD snap.

The pre-start/post-stop container hooks point to the LXD path and are
called when the container shuts down.

As the snapd paths include the revision of the daemon, a system which
got updated would then have a hook path pointing to a non-existing
binary.

This extra env variable allows us to point to the "current" path in the
snap, which is effectively a symlink to whichever is the latest binary.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>