devices: implement infiniband device detection
We use a hashmap where the key is the device name mapped to a given {P,V}F name
as seen on the host. This way we can later cheaply retrieve all info based on
the "parent" or "host_name" property of the relevant infiniband device.

Infiniband cards additionally expose a set of character devices that are used
for direct memory access. This includes uberb*, umad*, etc. Some of these
character devices are individual to each virtual device and some are per port
meaning that multiple devices share the same character device. For example,
uverb devices are shared between different virtual devices. LXD will
opportunistically add these devices if they are detected. The code assumes that
all infiniband devices are located in "/dev/infiniband/*". Custom udev rules
are thus not supported. Users that use custom infiniband character device paths
and names should pass them explictily via the unix-char device type LXD
exposes if they need them. This a) keeps the code performant and b) guarantees
that the device name is predictable and guarantees unique encoding when
creating the per-container device.

Closes #3983.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>