migrate: implement pre-copy migration
This uses the CRIU's pre-dump support to implement pre-copy migration in
LXD. To implement this on the CRIU side, CRIU uses the soft-dirty bit in
the pagemap:

 https://www.kernel.org/doc/Documentation/vm/soft-dirty.txt
 https://criu.org/Memory_changes_tracking

As long as there is no feature detection in LXC the pre-copy migration
needs to be enabled explicitly as not all architecture/kernel/criu
combinations support dirty pages tracking.

To enable pre-copy migration in LXD the following parameter needs to be set
to 'true':

 migration.incremental.memory

Currently it defaults to 'false'. As LXD does not yet know how at which
point it has done enough pre-copying the following parameter exists:

 migration.incremental.memory.iterations

This defaults to '10'. This means that LXD will do 10 pre-dumps before
doing the final dump.

Thanks to protobuf the whole pre-copy migration should be transparent
enough, that if the receiving side does not know anything about pre-copy
migration (and especially the additional rsync transfers) it just does
the non-optimized migration even if 'migration.incremental.memory' is set
to 'true'.

The following things will be improved in follow up commits:

 * read out CRIU's 'stats-dump' file to know how many pages are
   part of the previous dump
 * make pre-copy migration the default if LXC tells us the
   architecture/kernel/criu combination supports it
 * update documentation
 * update test cases

Signed-off-by: Adrian Reber <areber@redhat.com>
Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>