migration: prepare for pre-copy migration (part 2)
In addition to the previous pre-copy migration prepare commit which
changed the parameters of the Migrate() function this extends the
protocol between migration source and destination to detect if
pre-copy migration should be used.

Currently pre-copy migration defaults to off and can be enabled by
setting 'migration.incremental.memory' to true. If it is 'true' and the
migration destination side also acknowledges that it supports pre-copy
migration the variable 'use_pre_dumps' is set to 'true' and the
following commits can use this variable to use pre-copy migration or not
in the upcoming patches.

Signed-off-by: Adrian Reber <areber@redhat.com>
Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>