migrate: finish pre-dumping if threshold reached
This adds support to stop the pre-dumping if a certain threshold has
been reached. Instead of blindly relying on the 'max' iterations this
commit makes it possible to finish the pre-dumping if enough memory
has already been migrated using the pre-copy migration optimization.

The configuration option is:

 migration.incremental.memory.goal (default to 70%)

If the current pre-dump more pages (%) migrated using pre-dumping
than defined in above variable the pre-copy migration is finished even
if the 'max' iterations has not been reached.

The lower the number the sooner the pre-copy migration will finish, but
the container downtime during migration will be longer as a result of
the time required to transfer the memory pages.

The higher the number is, the shorter the container downtime during
migration.

Signed-off-by: Adrian Reber <areber@redhat.com>
Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>