console: adapt to new liblxc changes
Not too long ago we merged

https://github.com/lxc/lxc/pull/1925

which adds the new configuration keys

- lxc.console.buffer.size
- lxc.console.buffer.logfile
- lxc.console.rotate

LXD will always set

lxc.console.buffer = auto

which allocates a 128kB ringbuffer for each container and

lxc.console.buffer.logfile=/var/log/lxd/<container-name>/console.log

which is an on-disk representation of the console ringbuffer. liblxc will
always dump the in-memory ringbuffer when a container stops so users can
request the last contents of the ringbuffer if a container fails to start. When
a container is running and a request is sent to the retrieve the current
in-memory ringbuffer contents liblxc will also automatically dump the current
contents to disk.

As of this commit LXD will ignore lxc.console.logfile i.e. it cannot be
retrieved via the API. We still allow users to set this file via raw.lxc but
the API does not allow interaction with it. Users can either request the
in-memory ringbuffer contents or its on-disk representation so there' really no
point in having this exposed in the API.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>