migration: remove obsolete TODO comment
Initially the idea was to check for pre-copy support also on the
receiving side. CRIU supports pre-copying for a 'very long' time already
and there is no special kernel support required to restore a pre-copy
checkpoint. If the sender wants to do pre-copy migration let's just do it.

If the receiving side does not include the code for pre-copy migration
the protobuf header will neither be filled with 'true' or 'false'
concerning the pre-copy support of the receiving side which will be
interpreted as 'false' on the side of the sender as the protocol to
handle multiple rsync transfers for each pre-copy iteration is only
available after the pre-copy support has been added.

Interestingly this is a rather long commit message for a single line
comment removal.

Signed-off-by: Adrian Reber <areber@redhat.com>