Fix regression in "lxd init" in interactive mode for clustering
The lxd init code used to get networks and storages via a dedicated
cluster API, which was filtering out managed networks. The code was
updated to use the regular GET /networks API instead, and to filter
out managed networks. However it was creating a slice of networks
assuming that no network would be skipped.

I tried to add an integration test for interactive mode, but filling
in the trust password is not straightforward. We'll probably want to
figure it out separately.

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>