devlxd: Add image download API
This introduces a new API to devlxd which allows downloading of images
stored on the host. It requires the images on the host to either be
public or be marked as cached and requires the full image fingerprint.

This API only gives access to the image blob, not any of its metadata as
that data may not be relevant for a container.

The intended use for this is to minimize network usage for those using
nested LXDs when the image used is already in the host's image store.
The nested LXD will still hit the image store to grab the current image
metadata, resolve aliases, ... but the image blob itself will be
downloaded from the host over devlxd.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>