Serialize reads to the cluster database
This change makes us use transactions also in test code, and the /internal/sql
API endpoint which was not doing that before.

It also drops concurrent calls in the GET /containers and cluster heartbeat
code, since at the moment they are hardly going to take advantage of
concurrency, as the nodes are going to serialize db reads anyways (and db reads
atm are a substantial part of the total time spent in handling an API request).

The lower-level change to actually serialize reads was committed partly in
go-grpc-sql and partly in dqlite.

This should mitigate #4548 for now.

Moving forward we should start optimizing dqlite to be faster (I believe there
are substantial gains to be made there) , and perhaps also change the LXD code
that interacts with the database to be more efficient (e.g. caching prepared
statements, not entering/exiting a transaction for every query, etc).

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>