Fix the storage_pool_id filter from the WHERE clause of StoragePoolsConfig
The filter was broken since it should have been "storage_pools_config.node_id=?"
instead.

The method has been renamed to StoragePoolsNodeConfig for extra clarity.

I changed the logic of NetworksConfig to perform node filtering as well (and
renamed it to NetworksNodeConfig), and dropped Go-land manual filtering in the
call sites.

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>