Allow identity mappings for unprivileged containers
Allow identity mapped uid/gid ranges in the raw.idmap config even if the
LXD_UNPRIVILEGED_ONLY environment variable is set.  Make sure that this
also prevents the root user in the parent namespace from being mapped
into the child namespace.

This is useful for mapping the uid of the user into the child namespace
so that they can still access files in their home directory.

Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>