proxy: handle full socket buffer
Usually we should wait for the child process somewhere here. But we cannot
really do this. The listener file descriptors are retrieved in the go runtime
but at that point we have already double-fork()ed to daemonize ourselves and so
we can't wait on the child anymore after we received the listener fds. On the
other hand, if we wait on the child here we wait on the child before the
receive. However, if we do this then we can end up in a situation where the
socket send buffer is full and we need to retrieve some file descriptors first
before we can go on sending more. But this won't be possible because we're
waiting before the call to receive the file descriptor in the go runtime.
Luckily, we can just rely on init doing it's job and reaping the zombie
process. So, technically unsatisfying but pragmatically correct.

Explanatory Anectdote:
The limit for sending fds lies somewhere in the hundreds. I've hit this limit
before when sending pty fds in LXC around. Once there was code in LXC that did
this in one send. So if a user requested 500 fds LXC would try to send 500 fds
from parent to child and it would fail with EINVAL. That's when I switched to
batch-based sending. This is a similar situation just that we fill the buffer
bit by bit.

Closes #4681.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>