lxd: Add built-in pprof server
This commit adds a new core.debug_address option (node specific) which
controls a built-in HTTP server to be used with the "pprof" tool or
directly from a web browser (/debug/pprof is the main endpoint).

This is a much more flexible option that can be live-enabled by our
users when running into a problem and replaces the old cpu-profile,
memory-profile and print-goroutines command line options.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>