Reduce the frequency of raft snapshots
We have been taking raft snapshots agressively for a while now, to mitigate

This change makes the frequency less aggressive. It shouldn't really make a
difference regarding #4485, since we take a raft snapshot at daemon shutdown
anyways, so the net effect is virtually the same (except in case of hard
crashes).

The snapshots trick will go away altogether in the new dqlite implementation.

Signed-off-by: Free Ekanayaka <free.ekanayaka@canonical.com>