lxd/networks: Improve dnsmasq leases cleanup
dnsmasq doesn't record the MAC address for IPv6 leases, so add logic to
detect that case and then cleanup base on the container name.

This isn't perfect as there's no guarantee that the reported name over
DHCP matches the container's LXD name, but there's really nothing else
to match against.

Closes #4635

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>