networks: Fix packet stats logic for containers
The new logic introduced in 3.3 was both slower due to having to do
multiple file operations and also wrong as forknet doesn't actually have
the container's /sys view.

The result was the expected list of interfaces and addresses but
statistics coming from the host's device, if a matching one existed.

Closes #4851

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>