idmap: convert uid from big to little endian
When I suggested the memcpy() between vfs_cap_data to vfs_ns_cap_data in order
to copy fscaps I forgot to mention that we need to ensure that all fields are
passed as 32 bit little endian integers. This commit makes sure to convert from
big to little endian when passing the uid down.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>